pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Visual Question and Answering"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10.3"
click = "^8.1.3"
torch = "^1.12.1"
transformers = "^4.21.3"
loguru = "^0.6.0"
Pillow = "^9.2.0"
pandas = "^1.4.4"
black = "^22.8.0"
opencv-python = "^4.6.0"
matplotlib = "^3.5.3"
torchvision = "^0.13.1"
tensorboard = "^2.10.0"
img2vec-pytorch = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

packages = [
    { include = "scripts" }
]

[tool.poetry.scripts]
capstone = "src.main:main"

